[IMP] account_payment_mandate: large code improvement detailed below.
partner_id on mandate is not a related field any more, it is a regular
field. When you create a mandate, you first select the partner and then
the bank account among the bank accounts of that partner.
Fixes issue https://github.com/OCA/bank-payment/issues/1473

2 datamodel changes:
- introduce 'final' in the list of states
- type has 2 possible values: 'recurrent' or 'oneoff' ('generic' has been removed)
These values are generic values, they should not be considered as SEPA-specific.
Advantage: it allows to remove fields on the mandate in the module account_payment_sepa_direct_debit, which improves usability and simplicity.

Add buttons on mandate to switch from 'valid' to 'final' (and go back in case of mistake).

Update the code base to replace filters on state == 'valid' by a filter state in ('valid', 'final')

Improve error messages

Re-organise form view of mandate on 2 columns. Improve list view. Update search view.

Update tests